[Admin][Products]
- add default images to products, blog and fix stripe payment bug

+++ src/app/admin-profile/product-management/product-actions/handle-accessories/handle-accessories.component.ts
@@ -22,6 +22,7 @@ import { ProductReeviews } from '../../product-models/product-reviews.model';
 import { Editor, Toolbar } from 'ngx-editor';
 import { LoadingService } from '../../../../loading-spinner/loading.service';
 import { Timestamp } from 'firebase/firestore';
+import { DefaultImageUrl } from '../../../default-image-url';
 
 @Component({
   selector: 'app-handle-accessories',
@@ -184,7 +185,7 @@ export class HandleAccessoriesComponent {
           // If not found, push the unique color and its imageUrl to the accessoryColors array
           this.accessoryColors.push({
             color: price.productColor,
-            imageUrl: price.productImage || '' // Use an empty string if no image
+            imageUrl: price.productImage || DefaultImageUrl.productUrl // Use an empty string if no image
           });
         }
       }
@@ -217,8 +218,8 @@ export class HandleAccessoriesComponent {
     dialogRef.afterClosed().subscribe((newColor: ProductColor) => {
       if (newColor) {
         // Ensure color product image is not null
-        if (newColor.imageUrl === null || newColor.imageUrl === undefined) {
-          newColor.imageUrl = "";
+        if (newColor.imageUrl === null || newColor.imageUrl === undefined || newColor.imageUrl === '') {
+          newColor.imageUrl = DefaultImageUrl.productUrl;
         }
 
         // Check if the color already exists in the array
@@ -252,8 +253,8 @@ export class HandleAccessoriesComponent {
       } else if (editedColor && typeof editedColor === 'object') {
         // If an edited color is returned, update the accessoryColors array
         // Ensure product image is not null
-        if (editedColor.imageUrl === null || editedColor.imageUrl === undefined) {
-          editedColor.imageUrl = "";
+        if (editedColor.imageUrl === null || editedColor.imageUrl === undefined || editedColor.imageUrl === '') {
+          editedColor.imageUrl = DefaultImageUrl.productUrl;
         }
 
         // Check if the color already exists in the array
@@ -304,7 +305,7 @@ export class HandleAccessoriesComponent {
   // handle unified checkbox value
   onUnifiedImageChange(isUnified: boolean) {
     if (isUnified) {
-      this.unifiedImageUrl = '';
+      this.unifiedImageUrl = DefaultImageUrl.productUrl;
       this.productPrices.forEach(price => {
         price.productImage = this.unifiedImageUrl
       });
@@ -330,7 +331,7 @@ export class HandleAccessoriesComponent {
   }
 
   removeImage() {
-    this.unifiedImageUrl = '';
+    this.unifiedImageUrl = DefaultImageUrl.productUrl;
     this.productPrices.forEach(price => {
       price.productImage = this.unifiedImageUrl
     });
@@ -355,7 +356,7 @@ export class HandleAccessoriesComponent {
       if (newPrice) {
         // Ensure product image is not null
         if (newPrice.productImage === null || newPrice.productImage === undefined) {
-          newPrice.productImage = "";
+          newPrice.productImage = DefaultImageUrl.productUrl;
         }
 
         // if new price is the default then change the old default value to false
@@ -416,7 +417,7 @@ export class HandleAccessoriesComponent {
         // If an edited price is returned, update the productPrices array
         // Ensure product image is not null
         if (editedPrice.productImage === null || editedPrice.productImage === undefined) {
-          editedPrice.productImage = "";
+          editedPrice.productImage = DefaultImageUrl.productUrl;
         }
 
         // if edited price is the default then change the old default value to false
@@ -611,7 +612,7 @@ export class HandleAccessoriesComponent {
         try {
           // Delete images from Firebase Storage
           const deleteImagePromises = this.productPrices.map(async (price: ProductPrice) => {
-            if (price.productImage) {
+            if (price.productImage && price.productImage !== DefaultImageUrl.productUrl) {
               try {
                 const fileRef = this.storage.refFromURL(price.productImage);
                 await fileRef.delete().toPromise();
