+++ src/app/admin-profile/product-management/product-actions/handle-food-supliments/handle-food-supliments.component.ts
@@ -21,6 +21,7 @@ import { AdminService } from '../../../admin.service';
 import { Editor, Toolbar } from 'ngx-editor';
 import { LoadingService } from '../../../../loading-spinner/loading.service';
 import { Timestamp } from 'firebase/firestore';
+import { DefaultImageUrl } from '../../../default-image-url';
 
 @Component({
   selector: 'app-handle-food-supliments',
@@ -510,7 +511,7 @@ export class HandleFoodSuplimentsComponent implements OnInit {
   // handle unified checkbox value
   onUnifiedImageChange(isUnified: boolean) {
     if (isUnified) {
-      this.unifiedImageUrl = '';
+      this.unifiedImageUrl = DefaultImageUrl.productUrl;
       this.productPrices.forEach(price => {
         price.productImage = this.unifiedImageUrl
       });
@@ -536,7 +537,7 @@ export class HandleFoodSuplimentsComponent implements OnInit {
   }
 
   removeImage() {
-    this.unifiedImageUrl = '';
+    this.unifiedImageUrl = DefaultImageUrl.productUrl;
     this.productPrices.forEach(price => {
       price.productImage = this.unifiedImageUrl
     });
@@ -559,8 +560,8 @@ export class HandleFoodSuplimentsComponent implements OnInit {
     dialogRef.afterClosed().subscribe((newPrice: ProductPrice) => {
       if (newPrice) {
         // Ensure product image is not null
-        if (newPrice.productImage === null || newPrice.productImage === undefined) {
-          newPrice.productImage = "";
+        if (newPrice.productImage === null || newPrice.productImage === undefined || newPrice.productImage === '') {
+          newPrice.productImage = DefaultImageUrl.productUrl;
         }
 
         // if new price is the default then change the old default value to false
@@ -613,8 +614,8 @@ export class HandleFoodSuplimentsComponent implements OnInit {
       } else if (editedPrice && typeof editedPrice === 'object') {
         // If an edited price is returned, update the productPrices array
         // Ensure product image is not null
-        if (editedPrice.productImage === null || editedPrice.productImage === undefined) {
-          editedPrice.productImage = "";
+        if (editedPrice.productImage === null || editedPrice.productImage === undefined || editedPrice.productImage === '') {
+          editedPrice.productImage = DefaultImageUrl.productUrl;
         }
 
         // if edited price is the default then change the old default value to false
@@ -959,7 +960,7 @@ export class HandleFoodSuplimentsComponent implements OnInit {
         try {
           // Delete images from Firebase Storage
           const deleteImagePromises = this.productPrices.map(async (price: ProductPrice) => {
-            if (price.productImage) {
+            if (price.productImage && price.productImage !== DefaultImageUrl.productUrl) {
               try {
                 const fileRef = this.storage.refFromURL(price.productImage);
                 await fileRef.delete().toPromise();
