+++ src/app/admin-profile/product-management/product-actions/handle-organic-food/handle-organic-food.component.ts
@@ -21,6 +21,7 @@ import { Editor, Toolbar } from 'ngx-editor';
 import { ProductReeviews } from '../../product-models/product-reviews.model';
 import { LoadingService } from '../../../../loading-spinner/loading.service';
 import { Timestamp } from 'firebase/firestore';
+import { DefaultImageUrl } from '../../../default-image-url';
 
 @Component({
   selector: 'app-handle-organic-food',
@@ -314,7 +315,7 @@ export class HandleOrganicFoodComponent {
   // handle unified checkbox value
   onUnifiedImageChange(isUnified: boolean) {
     if (isUnified) {
-      this.unifiedImageUrl = '';
+      this.unifiedImageUrl = DefaultImageUrl.productUrl;
       this.productPrices.forEach(price => {
         price.productImage = this.unifiedImageUrl
       });
@@ -340,7 +341,7 @@ export class HandleOrganicFoodComponent {
   }
 
   removeImage() {
-    this.unifiedImageUrl = '';
+    this.unifiedImageUrl = DefaultImageUrl.productUrl;
     this.productPrices.forEach(price => {
       price.productImage = this.unifiedImageUrl
     });
@@ -364,8 +365,8 @@ export class HandleOrganicFoodComponent {
     dialogRef.afterClosed().subscribe((newPrice: ProductPrice) => {
       if (newPrice) {
         // Ensure product image is not null
-        if (newPrice.productImage === null || newPrice.productImage === undefined) {
-          newPrice.productImage = "";
+        if (newPrice.productImage === null || newPrice.productImage === undefined || newPrice.productImage === '') {
+          newPrice.productImage = DefaultImageUrl.productUrl;
         }
 
         // if new price is the default then change the old default value to false
@@ -419,8 +420,8 @@ export class HandleOrganicFoodComponent {
       } else if (editedPrice && typeof editedPrice === 'object') {
         // If an edited price is returned, update the productPrices array
         // Ensure product image is not null
-        if (editedPrice.productImage === null || editedPrice.productImage === undefined) {
-          editedPrice.productImage = "";
+        if (editedPrice.productImage === null || editedPrice.productImage === undefined || editedPrice.productImage === '') {
+          editedPrice.productImage = DefaultImageUrl.productUrl;
         }
 
         // if edited price is the default then change the old default value to false
@@ -658,7 +659,7 @@ export class HandleOrganicFoodComponent {
         try {
           // Delete images from Firebase Storage
           const deleteImagePromises = this.productPrices.map(async (price: ProductPrice) => {
-            if (price.productImage) {
+            if (price.productImage && price.productImage !== DefaultImageUrl.productUrl) {
               try {
                 const fileRef = this.storage.refFromURL(price.productImage);
                 await fileRef.delete().toPromise();
